requirements.txt
pydantic~=2.9.2
puzpy~=0.2.6
lxml~=5.3.0
unidecode~=1.3.8
pytest~=8.3.3
openai~=1.86.0
python-dotenv~=1.0.1

# Web UI dependencies
flask~=3.1.0
flask-socketio~=5.4.1
flask-cors~=5.0.0
python-socketio~=5.12.1